🚚 Move SocketMessageType to shared
