+++ packages/scraper/src/types/index.ts
@@ -4,4 +4,3 @@ export type { NoticeScript } from "./NoticeScript";
 export type { Calendar } from "./Calendar";
 export type { CalendarScript } from "./CalendarScript";
 export type { CafeteriaScript } from "./CafeteriaScript";
-export { SocketMessageType } from "./Socket";
